Pagination ready and works, but need to update tests for GetAdList.

+++ SpeedyWheelsSales/src/SpeedyWheelsSales.Application/Features/Ad/Queries/GetAdList/DTOs/AdListCarDto.cs
@@ -7,7 +7,7 @@ public class AdListCarDto
     public string Model { get; set; }
     public decimal Price { get; set; }
     public int Mileage { get; set; }
-    public decimal EngineSize { get; set; }
+    public double EngineSize { get; set; }
     public string Vin { get; set; }
     public string Plates { get; set; }
     public DateTime ManufactureDate { get; set; }
